More informative error messages
When a character was found for which a check hasn't been defined, the previous behavior was to output (to standard out) that character with a message that it had been found but not defined. This was changed to add the hex notation for that character, so it was clear precisely what it was. (Simply printing the character was a problem when what was printed looked just like another character; case in point, "!" 0x0021 "EXCLAMATION MARK" and "ǃ" 0x01c3 "LATIN LETTER RETROFLEX CLICK")